backend/requirements.txt
flask
flask-restful
google-cloud-firestore
google-cloud-pubsub
gunicorn
marshmallow